* Add max and min post aggregators that can be used with the approxHistogramFold Aggregation
With the following extension:

http://druid.io/docs/latest/development/extensions-core/approximate-histograms.html

it is possible to calculate the the maximum and minimum values of the underlying approximate
histogram aggregator. This adds support to the druid grafana plugin so that those post
aggregators can be used to build displays of the data. This saves having to use an
additional max aggregator on ingestion if the approx histogram is already being
calculated.